refactor: remove unused tests
