+++ apps/ui/components/form/__tests__/ScollIntoViewWithFocus.test.tsx
@@ -1,9 +1,8 @@
 import { render } from "@testing-library/react";
 import React from "react";
-
 import ScrollIntoViewWithFocus from "../ScrollIntoViewWithFocus";
 
-it("should scroll to component when focused", async () => {
+test("should scroll to component when focused", async () => {
   const scrollIntoViewMock = jest.fn();
   window.HTMLElement.prototype.scrollIntoView = scrollIntoViewMock;
   render(
