+++ packages/common/package.json
@@ -47,7 +47,7 @@
   },
   "devDependencies": {
     "@apollo/client": "^3.9.7",
-    "@testing-library/jest-dom": "^6.2.0",
+    "@testing-library/jest-dom": "^6.4.6",
     "@testing-library/react": "^16.0.0",
     "@testing-library/user-event": "^14.5.2",
     "@types/jest": "^29.5.12",
